requirements.txt
Django~=4.2.5
celery~=5.3.4
djangorestframework~=3.14.0